+++ assets/css/app/components/_node_selection.scss
@@ -1,7 +1,7 @@
 // Node selection dropdown
 #node-selection {
   position: absolute;
-  top: -5.5rem;
+  top: -4.5rem;
   right: -0.75rem;
   padding-right: 0.75rem;
   padding-top: 0.75rem;
