+++ lib/phoenix/live_dashboard/live/menu_live.ex
@@ -1,25 +1,30 @@
 defmodule Phoenix.LiveDashboard.MenuLive do
   use Phoenix.LiveDashboard.Web, :live_view
 
+  @default_refresh 5
+  @supported_refresh [{"1s", 1}, {"2s", 2}, {"5s", 5}, {"15s", 15}, {"30s", 30}]
+
   @impl true
   def mount(_, %{"menu" => menu}, socket) do
-    socket = assign(socket, menu: menu, node: menu.node)
+    socket = assign(socket, menu: menu, node: menu.node, refresh: @default_refresh)
     socket = validate_nodes_or_redirect(socket)
 
     if connected?(socket) do
       :net_kernel.monitor_nodes(true, node_type: :all)
     end
 
-    {:ok, socket}
+    {:ok, init_schedule_refresh(socket)}
   end
 
   @impl true
   def render(assigns) do
     ~L"""
-    <%= maybe_active_live_redirect @socket, "Home", :home, @node %>
-    <%= maybe_enabled_live_redirect @socket, "Metrics", :metrics, @node %>
-    <%= maybe_enabled_live_redirect @socket, "Request Logger", :request_logger, @node %>
-    <%= maybe_active_live_redirect @socket, "Processes", :processes, @node %>
+    <nav id="menu-bar">
+      <%= maybe_active_live_redirect @socket, "Home", :home, @node %>
+      <%= maybe_enabled_live_redirect @socket, "Metrics", :metrics, @node %>
+      <%= maybe_enabled_live_redirect @socket, "Request Logger", :request_logger, @node %>
+      <%= maybe_active_live_redirect @socket, "Processes", :processes, @node %>
+    </nav>
 
     <form id="node-selection" phx-change="select_node" class="d-inline">
       <div class="input-group input-group-sm d-flex flex-column">
@@ -31,7 +36,30 @@ defmodule Phoenix.LiveDashboard.MenuLive do
         </select>
       </div>
     </form>
+
+    <div id="refresher">
+      <form phx-change="select_refresh">
+        <div class="input-group input-group-sm">
+          <%= if @menu.refresher? do %>
+            <div class="input-group-prepend">
+              <label class="input-group-text" for="refresh-interval-select">Update every</label>
+            </div>
+            <select name="refresh" class="custom-select" id="refresh-interval-select">
+              <%= options_for_select(refresh_options(), @refresh) %>
+            </select>
+          <% else %>
+            <div class="input-group-prepend">
+              <small class="input-group-text text-muted">Updates automatically</small>
+            </div>
+          <% end %>
+        </div>
+      </form>
+    </div>
     """
+  end
+
+  defp refresh_options() do
+    @supported_refresh
   end
 
   defp maybe_active_live_redirect(socket, text, action, node) do
@@ -67,6 +95,11 @@ defmodule Phoenix.LiveDashboard.MenuLive do
     {:noreply, validate_nodes_or_redirect(socket)}
   end
 
+  def handle_info(:refresh, socket) do
+    send(socket.root_pid, :refresh)
+    {:noreply, schedule_refresh(socket)}
+  end
+
   @impl true
   def handle_event("select_node", params, socket) do
     param_node = params["node"]
@@ -78,6 +111,27 @@ defmodule Phoenix.LiveDashboard.MenuLive do
     else
       {:noreply, redirect_to_current_node(socket)}
     end
+  end
+
+  def handle_event("select_refresh", params, socket) do
+    case Integer.parse(params["refresh"]) do
+      {refresh, ""} -> {:noreply, assign(socket, refresh: refresh)}
+      _ -> {:noreply, socket}
+    end
+  end
+
+  ## Refresh helpers
+
+  defp init_schedule_refresh(socket) do
+    if connected?(socket) and socket.assigns.menu.refresher? do
+      schedule_refresh(socket)
+    else
+      assign(socket, timer: nil)
+    end
+  end
+
+  defp schedule_refresh(socket) do
+    assign(socket, timer: Process.send_after(self(), :refresh, socket.assigns.refresh * 1000))
   end
 
   ## Node helpers
